requirements.txt
groq==0.4.2
streamlit==1.39.0
python-dotenv==1.0.0
yt-dlp==2024.12.23
